work on statemachine notebook
